Spatial Cropping and Masking for PISCO Rasters

Crops and masks a PISCO terra::SpatRaster using a vector polygon (sf or SpatVector), an sf::st_bbox object, or a spatial bounding box numeric vector.

Usage

pisco_clip(x, mask, crop_only = FALSE)

Arguments

x

A terra::SpatRaster object.

mask

An sf object, terra::SpatVector, sf::st_bbox, terra::SpatExtent, or numeric bounding box (c(xmin, ymin, xmax, ymax)).

crop_only

Logical. If TRUE, only crops to bounding box without masking values outside polygon. Default is FALSE.

Value

A cropped (and optionally masked) terra::SpatRaster.

Examples

if (FALSE) { # \dontrun{
r <- pisco_read("monthly")
# Crop by bounding box (e.g. Lima / Rimac basin)
r_sub <- pisco_clip(r, mask = c(-77.5, -12.5, -76.0, -11.5))

# Crop and mask using an sf polygon
# r_basin <- pisco_clip(r, mask = basin_sf)
} # }